98 fibonacci

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -5,16 +5,16 @@
  */
 int main(void)
 {
-unsigned long int a, b, c, i;
+unsigned long long int a, b, c, i;
 a = 0;
 b = 1;
 c = 1;
 for (i = 1; i <= 98; i++)
 {
 if (i != 98)
-printf("%lu, ", c);
+printf("%llu, ", c);
 else
-printf("%lu", c);
+printf("%llu", c);
 a = b;
 b = c;
 c = a + b;
